// src/erp-ui/projects/api/src/lib/model/booking-dto.ts
/**
 * Fachschaften-ERP.Api | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CashIncomeSource } from './cash-income-source';
import { TaxArea } from './tax-area';


export interface BookingDto { 
    categoryName?: string | null;
    cashBookName?: string | null;
    id: string;
    date: string;
    amount: number;
    description: string;
    taxArea?: TaxArea | null;
    categoryId?: string | null;
    cashBookId?: string | null;
    receiptUrl?: string | null;
    source?: CashIncomeSource | null;
    transferGroupId?: string | null;
}