// src/erp-ui/projects/api/src/lib/model/tax-area.ts
/**
 * Fachschaften-ERP.Api | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum TaxArea {
    Ideell = 'Ideell',
    Zweckbetrieb = 'Zweckbetrieb',
    Wirtschaftsbetrieb = 'Wirtschaftsbetrieb'
}